Added msub and andfc/andcf to LLE, also fixed compiler warning.
git-svn-id: https://dolphin-emu.googlecode.com/svn/trunk@2844 8ced0084-cf51-0410-be5f-012b33b47a6e

+++ Source/Plugins/Plugin_DSP_LLE-testing/Src/disassemble.cpp
@@ -219,7 +219,7 @@ char* gd_dis_params(gd_globals_t* gdg, DSPOPCTemplate* opc, u16 op1, u16 op2, ch
 			if (opc->params[j].size != 2)
 			{
 				if (opc->params[j].mask == 0x007f) // LSL, LSR, ASL, ASR
-					sprintf(buf, "#%d", val < 64 ? val : -(0x80-val)); // I know compiler complains, but this is in gcdsptool...
+					sprintf(buf, "#%d", val < 64 ? val : -(0x80-(s32)val));
 				else
 					sprintf(buf, "#0x%02x", val);
 			}
